+++ src/main/java/com/wellnesswave/WellnessWave/Repository/AppointmentsRepository.java
@@ -3,5 +3,5 @@ package com.wellnesswave.WellnessWave.Repository;
 import com.wellnesswave.WellnessWave.Entities.Appointments;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentsRepository extends JpaRepository<Appointments, Long> {
+public interface AppointmentsRepository extends JpaRepository<Appointments, Integer> {
 }
